+++ src/components/hook/bizPost/useBizPostDetailQuery.js
@@ -5,6 +5,6 @@ export const useBizPostDetailQuery = (params) => {
   return useQuery({
     queryKey: ["bizPostDetail"],
     queryFn: () => bizPostDetailApi(params),
-    staleTime: 1000 * 60,
+    // staleTime: 1000 * 60,
   });
 };
