+++ src/components/hook/bizPost/useBizPostListSearchQuery.js
@@ -5,6 +5,6 @@ export const useBizPostListSearchQuery = (cPage, injectedValue,userType) => {
   return useQuery({
     queryKey: ["HirePost", injectedValue, cPage],
     queryFn: () => bizPostListSearchApi(injectedValue.value, cPage.value,userType),
-    staleTime: 1000 * 60,
+    // staleTime: 1000 * 60,
   });
 };
