+++ src/components/hook/bizPost/useBizPostStatusUpdateMutation.js
@@ -1,19 +1,23 @@
 import { useMutation, useQueryClient } from "@tanstack/vue-query";
 import { useRouter } from "vue-router";
-import { bizPostDetailDeleteApi } from "../../../api/bizPost/bizPostDetailDeleteApi";
+import { bizPostStatusUptateApi } from "../../../api/bizPost/bizPostStatusUptateApi";
 
-export const useBizPostDetailDeleteMutation = (postDetail, bizDetail) => {
+export const useBizPostStatusUpdateMutation = () => {
   const queryClient = useQueryClient();
   const router = useRouter();
   return useMutation({
-    mutationFn: () => bizPostDetailDeleteApi(postDetail.value.postIdx, bizDetail.value.bizIdx),
-    mutationKey: ["bizPostDelete"],
-    onSuccess: () => {
-      alert("삭제 처리되었습니다.");
-      router.go(-1);
+    mutationFn: ({ postIdx, status }) => bizPostStatusUptateApi(postIdx, status),
+    mutationKey: ["bizPostStatusUptate"],
+    onSuccess: (data, { status }) => {
+      alert("처리되었습니다.");
       queryClient.invalidateQueries({
         queryKey: ["bizPostDetail"],
       });
+      if (status == "승인") {
+        router.push({ name: "managePost" });
+      } else if (status == "불허") {
+        router.push({ name: "managePostApproval" });
+      }
     },
   });
 };
